% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{download_bioclim_bio}
\alias{download_bioclim_bio}
\title{Downloads the WorldClim bioclimatic variables at 10 arc minute resolution.}
\usage{
download_bioclim_bio()
}
\description{
Downloads the WorldClim bioclimatic variables at 10 arc minute resolution.
}
\examples{
download_bioclim_bio()
}
