% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{modify_south_america}
\alias{modify_south_america}
\title{Crops or masks the map region covering South America.}
\usage{
modify_south_america(dat, modify = "Crop")
}
\arguments{
\item{dat}{a S4 class 'SpatRaster' that is subsetted to retain the
relevant bioclimatic variables}

\item{modify}{indicates whether to crop or mask South America}
}
\description{
Crops or masks the map region covering South America.
}
\examples{
\dontrun{dat <- download_bioclim_bio()[[1]]
mod <- modify_south_america(dat)
terra::plot(mod)}
}
